remember that we dropped frames
previously, pressing R did not reset the movie from the beginning like
it should. instead, the frame remained still until the play head
reached the time it was before we pressed R. This is because I forgot
to set the firstFrameDropped flag to true, and thus getPlayFrame
thought that the oldest remaining frame is the first frame of the
video.

Unfortunately, the new behaviour is worse than before: we display a
loading page, and the application becomes unresponsive!