merge the loading and resetting threads
The reason the app was forever stuck loading is that both the conditions
for loading a new frame and for resetting became satisfied, but the
loading thread kept winning the race, starving the resetting thread.
This was incorrect: when the resetting condition becomes satisfied, it
should be prioritized, as loading more frames at this point is wasted
work.

I'm still not sure why the app became unresponsive though; since the UI
and the loading thread run in different threads, shouldn't the UI thread
always be able to respond to the keyboard even if the loading thread is
pegging the CPU? Maybe ffmpeg-light's use of FFI blocks all the green
threads, so I need to run the loading thread in an OS thread?