consolidate acid-state data
Normally, acid-state records all the transactions and replays them upon
startup. This takes time and space proportional to the number of
transactions. By saving a snapshot, acid-state will restore from the
snapshot, which takes time proportional to the length of the video, but
space proportional to the length of the video times the number of
snapshots. By archiving old snapshots and deleting the archive, we
finally get time and space proportional to the length of the video.