[breaking] store Rel instead of Pos
Pos was in pixels, Rel is a fraction between 0 and 1. All existing
videofile.face-trace folders are no longer supported. At this point in
development, only one such folder exists, so I didn't bother with a
migration plan.

The change was made in order to simplify the scaling logic. The stream
is now responsible for converting Coords to Rels, and it can now do so
with only the dimensions of the _scaled_ video. Previously, to convert a
Coord to a Pos, it was necessary to keep track of a lot more
information.