switch to simpler ffmpeg-light API
I previously forked ffmpeg-light in order to add the
streamPixelAspectRatio function in the AvStreamT monad. Now that I am
getting ready to contribute that change back upstream, I realize that
this isn't idiomatic for this codebase, so I have changed that fork to
expose a getAspectRatio in the IO monad instead.