[broken] vendor gloss-juicy
Movies have a lot of frames, so I need to disable caching for Pictures.
Unfortunately, gloss-juicy always enables caching, and gloss no longer
exposes the guts which would make it configurable to disable caching.

I could ask gloss-juicy to expose a variant of their API which supports
disabling caching, but that would either require adding a ton of new
helper functions or making a backwards-incompatible change for a niche
feature which few people are going to use. And since issues are disabled
on gloss-juicy's github, I can't start a discussion asking if a PR with
either of those approaches has a chance to get accepted.

so I'll just edit a local copy of juicy-pixels.

does not compile yet